+++ halotools/mock_observables/two_point_clustering/s_mu_tpcf.py
@@ -7,8 +7,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 import numpy as np
 
 from .clustering_helpers import (process_optional_input_sample2,
-    downsample_inputs_exceeding_max_sample_size, verify_tpcf_estimator,
-    tpcf_estimator_dd_dr_rr_requirements)
+    verify_tpcf_estimator, tpcf_estimator_dd_dr_rr_requirements)
 from ..mock_observables_helpers import (enforce_sample_has_correct_shape,
     get_separation_bins_array, get_line_of_sight_bins_array, get_period, get_num_threads)
 from ..pair_counters.mesh_helpers import _enforce_maximum_search_length
@@ -24,7 +23,7 @@ np.seterr(divide='ignore', invalid='ignore')  # ignore divide by zero in e.g. DD
 
 def s_mu_tpcf(sample1, s_bins, mu_bins, sample2=None, randoms=None,
         period=None, do_auto=True, do_cross=True, estimator='Natural',
-        num_threads=1, max_sample_size=int(1e6), approx_cell1_size=None,
+        num_threads=1, approx_cell1_size=None,
         approx_cell2_size=None, approx_cellran_size=None, seed=None):
     r"""
     Calculate the redshift space correlation function, :math:`\xi(s, \mu)`
@@ -101,12 +100,6 @@ def s_mu_tpcf(sample1, s_bins, mu_bins, sample2=None, randoms=None,
         calculation, in which case a multiprocessing Pool object will
         never be instantiated. A string 'max' may be used to indicate that
         the pair counters should use all available cores on the machine.
-
-    max_sample_size : int, optional
-        Defines maximum size of the sample that will be passed to the pair counter.
-        If sample size exeeds max_sample_size,
-        the sample will be randomly down-sampled such that the subsample
-        is equal to ``max_sample_size``. Default value is 1e6.
 
     approx_cell1_size : array_like, optional
         Length-3 array serving as a guess for the optimal manner by how points
@@ -208,7 +201,7 @@ def s_mu_tpcf(sample1, s_bins, mu_bins, sample2=None, randoms=None,
 
     # process arguments
     function_args = (sample1, s_bins, mu_bins, sample2, randoms, period,
-        do_auto, do_cross, estimator, num_threads, max_sample_size,
+        do_auto, do_cross, estimator, num_threads,
         approx_cell1_size, approx_cell2_size, approx_cellran_size, seed)
 
     sample1, s_bins, mu_bins, sample2, randoms, period, do_auto, do_cross, num_threads,\
@@ -265,7 +258,7 @@ def spherical_sector_volume(s, mu):
     Note that the extra factor of 2 is to get the reflection.
     """
     theta = np.arccos(mu)
-    
+
     vol = (2.0*np.pi/3.0) * np.outer((s**3.0), (1.0-np.cos(theta)))*2.0
     return vol
 
@@ -388,7 +381,7 @@ def pair_counts(sample1, sample2, s_bins, mu_bins, period,
 
 
 def _s_mu_tpcf_process_args(sample1, s_bins, mu_bins, sample2, randoms,
-        period, do_auto, do_cross, estimator, num_threads, max_sample_size,
+        period, do_auto, do_cross, estimator, num_threads,
         approx_cell1_size, approx_cell2_size, approx_cellran_size, seed):
     """
     Private method to do bounds-checking on the arguments passed to
@@ -402,9 +395,6 @@ def _s_mu_tpcf_process_args(sample1, s_bins, mu_bins, sample2, randoms,
 
     if randoms is not None:
         randoms = np.atleast_1d(randoms)
-
-    sample1, sample2 = downsample_inputs_exceeding_max_sample_size(
-        sample1, sample2, _sample1_is_sample2, max_sample_size, seed=seed)
 
     # process radial bins
     s_bins = get_separation_bins_array(s_bins)