+++ halotools/mock_observables/two_point_clustering/tests/test_s_mu_tpcf.py
@@ -25,8 +25,7 @@ def test_s_mu_tpcf_auto_nonperiodic():
     mu_bins = np.linspace(0, 1.0, 5)
 
     result_1 = s_mu_tpcf(sample1, s_bins, mu_bins, sample2=None,
-        randoms=randoms, period=None,
-        max_sample_size=int(1e4), estimator='Natural')
+        randoms=randoms, period=None, estimator='Natural')
 
     assert result_1.ndim == 2, "correlation function returned has wrong dimension."
 
@@ -43,8 +42,7 @@ def test_s_mu_tpcf_auto_periodic():
     mu_bins = np.linspace(0, 1.0, 10)
 
     result_1 = s_mu_tpcf(sample1, s_bins, mu_bins, sample2=None,
-        randoms=None, period=period,
-        max_sample_size=int(1e4), estimator='Natural')
+        randoms=None, period=period, estimator='Natural')
 
     assert result_1.ndim == 2, "correlation function returned has wrong dimension."